This now works on Windows
Now the project can be compiled and run on Windows. Just double-click on "make.cmd" to set the Environment variables, let it fail the first time, then double-click again to compile it. After it is compiled, double-click "3dcollaborativegame.exe" to run the game. Use WASD to move along the x y axies (+-x and +- y, NOT forward and backward) and Q and E to move up and down. The mouse controls where you are looking.