Fixed issue of needing to run make file twice
There was an issue of needing to set the PATH environment variable on Windows. This force the user to run the make file twice to compile. This has now been fixed by no longer changing the environment variables. Now the user only needs to run the make file once to compile. In addtion, there is a new file called make_clean that removes the .o and .exe files to prepare for pushing changes to the repository.