The program now runs in 64 bit
There have been many folders containing libraries that only work with 32 bit architecture in the repository. MinGW (the compiler we used to use) was a 32 bit compiler that could only make 32 bit programs. We have now made the swtich over to mingw64, which is able to compilie in 64 bit. Because of this switch, we no longer need any of the 32 bit libraries. Thus, all 32 bit only files have been removed.