Universal Binary JSON parser is mostly done
It can read in nearly any UBJ file without having to load the entire file all at once. The next steps are to change the graphics engine to use the UBJ format rather than the current custom one, add the ability to write to UBJ files, and add the more specialized features of UBJ such as the compact container formatting.